:bug: Fixes segment delimiter parsing issue
The parser would treat the segment delimiter as a signal that there is another
segment to parse when in fact this delimiter indicates _the end_ of the current
segment.

This was causing the long-standing problem where a message would need to have
trailing whitespace trimmed before being parsed and emitted messages would lack
the trailing segment delimiter.

For the sake of robustness, the parser doesn't require the trailing
segment delimiter on the last segment; if you trim the whitespace on
your message before parsing it will still parse out correctly.